// projects/gp-route-tracker/src/lib/interfaces/locationSearch.model.ts
    export interface LocationResult {
        place_id: string;
        licence: string;
        osm_type: string;
        osm_id: string;
        boundingbox: string[];
        lat: string;
        lon: string;
        display_name: string;
        class: string;
        type: string;
        importance: number;
        icon: string;
    }